Remove IndexDistance constraint on Changeset and Edit
This removes the constraint on Collection.IndexDistance on Changeset and
Edit. Conversions between integer types are a little annoying in Swift,
especially in Swift 3. Before the revamp of the integer protocols in
[SE-0104], we have to go through `IntMax` to convert between
integer types. We should be able to remove the `toIntMax()` calls when
we drop Swift 3 support.

More importantly, while this code currently compiles, it crashes with
this runtime error when running the unit tests:

GenericCache(0x114050798): cyclic metadata dependency detected, aborting

The same crash happens in Xcode 8.3.2 and Xcode 9b4 (Swift 3 mode; I
don't know if it was fixed in Swift 4).

Related bugs:
- [SR-4383]
- [SR-4732]
- [SR-5068]

The crash has something to do with nested generic type metadata. As
suggested in one of the issues on JIRA, the crash goes away if we
make the enum inside Edit indirect:

```swift
public struct Edit<C: Collection> where C.Iterator.Element: Equatable {
    ...

	indirect public enum Operation {
        ...
	}
    ...
```

I don't know if this is supposed to be a final fix or just a workaround
and this issue will eventually be fixed.

[SE-0104]: https://github.com/apple/swift-evolution/blob/master/proposals/0104-improved-integers.md
[SR-4383]: https://bugs.swift.org/browse/SR-4383
[SR-4732]: https://bugs.swift.org/browse/SR-4732
[SR-5068]: https://bugs.swift.org/browse/SR-5086