Spell out collection type declarations
It's easy to see that Array<Edit<Array<URL>>> is an array of edits of an
array of URLs, whereas [Edit<[URL]>] becomes harder to parse because one
has to distinguish the similar looking square and angle brackets.

Personally I also have a tendency to ignore the subtle square brackets
among the angle brackets and sometimes miss it's an Array (or could it be a Set?).

Spelling out the Collection types makes the human parsing logical,
consistent and clear.