Make Edit a subtype of Changeset
An edit is expressed with respect to a collection. It doesn't make sense
to have a stand-alone edit. That would be an edit... into what? To
express this relationship clearly, Edit is now a subtype of Changeset.

As a consequence, Edit no longer has to copy the generic type
specifications from Changeset. It merely inherits them thereof.

All changes in this commit reflect this one, rather important, change.